+++ src/js/App/Header/InsightsAbout.js
@@ -60,7 +60,7 @@ class InsightsAbout extends Component {
                 { name: 'Dashboard', path: `apps/dashboard/app.info.json`, version: 'N/A' },
                 { name: 'Inventory', path: `apps/inventory/app.info.json`, version: 'N/A' },
                 { name: 'Remediations', path: `apps/remediations/app.info.json`, version: 'N/A' },
-                { name: 'Vulnerabilities', path: 'apps/vulnerability/app.info.json', version: 'N/A' },
+                { name: 'Vulnerability', path: 'apps/vulnerability/app.info.json', version: 'N/A' },
                 { name: 'Compliance', path: 'apps/compliance/app.info.json', version: 'N/A' },
                 { name: 'Cost Management', path: 'apps/cost-management/app.info.json', version: 'N/A' },
                 { name: 'Insights', path: 'apps/insights/app.info.json', version: 'N/A' }
