+++ src/js/vulnerabilityStyles.js
